Refactor wasm message handlers
