+++ x/wasm/internal/keeper/recurse_test.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	"encoding/json"
+	"github.com/CosmWasm/wasmd/x/wasm/internal/keeper/wasmtesting"
 	"github.com/CosmWasm/wasmd/x/wasm/internal/types"
 	"testing"
 
@@ -40,14 +41,13 @@ var totalWasmQueryCounter int
 func initRecurseContract(t *testing.T) (contract sdk.AccAddress, creator sdk.AccAddress, ctx sdk.Context, keeper *Keeper) {
 	// we do one basic setup before all test cases (which are read-only and don't change state)
 	var realWasmQuerier func(ctx sdk.Context, request *wasmvmtypes.WasmQuery) ([]byte, error)
-	countingQuerier := &QueryPlugins{
-		Wasm: func(ctx sdk.Context, request *wasmvmtypes.WasmQuery) ([]byte, error) {
+	countingQuerier := &wasmtesting.MockQueryHandler{
+		HandleQueryFn: func(ctx sdk.Context, request wasmvmtypes.QueryRequest, caller sdk.AccAddress) ([]byte, error) {
 			totalWasmQueryCounter++
-			return realWasmQuerier(ctx, request)
-		},
-	}
+			return realWasmQuerier(ctx, request.Wasm)
+		}}
 
-	ctx, keepers := CreateTestInput(t, false, SupportedFeatures, nil, countingQuerier)
+	ctx, keepers := CreateTestInput(t, false, SupportedFeatures, WithQueryHandler(countingQuerier))
 	keeper = keepers.WasmKeeper
 	realWasmQuerier = WasmQuerier(keeper)
