+++ x/wasm/internal/keeper/wasmtesting/query_handler.go
@@ -1,0 +1,17 @@
+package wasmtesting
+
+import (
+	wasmvmtypes "github.com/CosmWasm/wasmvm/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+type MockQueryHandler struct {
+	HandleQueryFn func(ctx sdk.Context, request wasmvmtypes.QueryRequest, caller sdk.AccAddress) ([]byte, error)
+}
+
+func (m *MockQueryHandler) HandleQuery(ctx sdk.Context, caller sdk.AccAddress, request wasmvmtypes.QueryRequest) ([]byte, error) {
+	if m.HandleQueryFn == nil {
+		panic("not expected to be called")
+	}
+	return m.HandleQueryFn(ctx, request, caller)
+}
